Prune Geth to components necessary for EVM
cardinal-evm is intended to just provide an EVM library drived from
geth. This commit starts with Geth and prunes out everything not needed
for the evm, shifting some dependencies to cardinal-types.

This is a long way from functional. We'll need a StateDB implementation
to cache and grab data from storage, but this initial commmit prunes out
the parts of Geth not used by the vm package.