Add state copying
Features like estimate gas require the ability to copy state.

This required some rework of the journal, to ensure that a revert
on a copied state would roll back the copy instead of the original.
I don't think there are any practical cases where a copy would get
rolled back beyond the point where it was copied from, but this
should make it possible.