Pipelined EVM implementation
CPUs regularly use instruction pipelining: https://en.wikipedia.org/wiki/Instruction_pipelining

This allows them to have multiple operations executing in parallel when
one instruction does not rely on the output of the immediately preceding
instruction.

This commit experiments with introducing a form of pipelining to the EVM.
It shifts the stack from being uint256 values to essentially being promises
for uint256 values. Mathematical operations on these promises produce new
promises, allowing these operations to happen in the background where
appropriate, and blocking only when the outcome of an operation is dependent
on the resolved value of a stack value and cannot be deferred further.

The hope is that allowing certain types of operations (such as state lookups)
to happen in the background while operations that do not directly rely on
the value of the state lookup can continue operating, only blocking when
the value from the state lookup becomes a crucial piece of information.