Incorporate several upstream EVM optimizations
* #24048
* #24085
* #24026
* #24031
* #24040
* #23970
* #23952
* #23974
* #23977
* #23967
* #24066

Most of these seem like fairly straighforward optimizations, eg.
doing checks inside of opcodes instead of at the interpreter level
where they must be done for every operation. #24120 is a bit opaque,
and claims to take advantage of amd64 CPU level optimizations, but
I wouldn't be surprised if the Geth team has secretly snuck in a
security fix disguised as an optimization.