Add producer broadcast abstraction
Since different chains will have some differences in the data they
need to broadcast, and we don't want to hardcode all of the differences
into a single plugin, I've allowed the Cardinal PluGeth plugin to look
for other plugins and allow them to make adjustments to the weight,
updates, and deletes of a given block.

Immediately, we add a merge plugin to adjust the weights. In the near
future we'll probably add a polygon plugin to deal with Bor receipts,
for example.