Add weight gauge to merge plugin
Leading up to the merge we want to be able to monitor the total difficulty
to evaluate proxmity to the merge. Unfortunately, our metrics use int64
and the terminal total difficulty does not fit into an int64.

To address this, we are dividing the total difficulty by 10**16, so the value
we get is in a range that will increase by 1-2 with each block and we can
monitor the metric approaching ttd // 10**16.