Use resumption time from block
If no resumption time is specified, get the block timestamp and
use that to retrieve offsets.

Stored offsets can be unreliable, particularly when a syncing process
gets disrupted mid-sync and the node needs to resume. Block timestamps
tend to be more reliable, though it's still possible to override with
the --resumption.ts flag.

If, for some reason, the header cannot be loaded, it will fall back
to using the resumption token from cardinal storage.