Reset log level in polygon plugin
Bor is doing something weird with respect to log level. It seems to
have shifted log levels by one, so that 3 = warning when it says 3 = info.
Additionally, it seems to be setting the log level back to 3 regardless
of what the flag says.

This update will use the debug_verbosity RPC method to set it to its
intended value based on the flag.