Check tracked prefixes before adding to the database
Not all messages sent over cardinal streams are of interest to
Cardinal-EVM. For example logs, receipts, and transactions are needed
in Flume, but ignored by Cardinal-EVM.

With the Kafka Cardinal-Streams consumer, only tracked prefixes are
delivered, but with the Websockets consumer all prefixes are sent
across the wire and a part of the delivered PendingBatch. This means
that if we're syncing via websockets, we might be writing a lot of
additional information to disk that Cardinal-EVM will never use.

By checking keys against the set of tracked prefixes, we can ensure
that only useful information is recorded to the database.