Add websocket initialization for Cardinal
Historically, Cardinal has had a tedious multi-step process for
exporting data from Geth and importing it into a Cardinal database.
This has gotten us by, and despite the tedium actually has some benefits
(eg. visibility into the ETL process, ability to change things on
the fly), but is not something the average open-source user of Cardinal
will need. To simplify, this commit makes it so you can initialize
Cardinal's database over websockets.

To run, start the master with:

```
./geth [OPTIONS] -- --cardinal.init.enabled
```

```
./cardinal-rpc-url --init.url=wss://localhost:8546 cardinal-config.yaml
```

Note that this will cause the master at localhost:8546 to stop processing
blocks while the database is initialized, so you don't want to run with
--cardinal.init.enabled for normal operation.

Right now the process isn't very resilient - if there is a websockets error
the whole process must be restarted. Perhaps that will be addressed in the
future, but as this should be a relatively rare step it's not a high priority.