Use state trie as fallback for unavailable storage
Maybe this ought to be refactored into blockupdates, but for current
testing purposes it should work here.

If we can't get state data for a block from the snapshotter, this will
now try the state trie. It will probably only work for archive nodes,
because it requires two blocks in a row, and normally the state trie
only has two consecutive blocks for the same range of blocks as the
snapshotter.