Pull ChainID from schema instead of InitializeNode
UpdateStreamsSchema has historically only worked if plugins were
loaded in the right order, which is dependent on naming conventions.
Relying on plugin naming conventions for key functionality is not
ideal.

Instead of relying on the ChainID set in InitializeNode (which relies
on InitializeNode executing before UpdateStreamsSchema, which is unreliable
because the producer plugin calls UpdateStreamsSchema from its own
InitializeNode), this searches the schema for a known key, extracts the
chain id, and uses that chainid for schema updates.