Fix panic in Transient Storage
Copies of the statedb were getting nil transient storage maps, which
was causing a panic, which obviously we don't want.

It was compounded by not having upgraded Cardinal RPC to a version which
releases the call semaphore on panics. Once we exceed the concurrency
limit, servers would lock up entirely