Fix RANDOM opcode on non-merged post-cancun chains
On chains that are not managed by a Beacon chain, the RANDOM opcode
maps to opDifficulty, not opRandom. We handled this properly when
Polygon and ETC merged Shanghai, but overlooked it when moving on to
cancun.

This chain updates the JumpTable on post-shanghai networks that are
not Merged networks, setting the random opcode back to opDifficulty.