Initial simulatev1 work
This makes several initial changes to support implementing simulatev1.

This code won't compile. There's a whole bunch of TODOs, types that
need to be implemented, VM updates that need to be done, tests that
need to be ported over from Geth, etc, but I wanted to get the initial
work somewhere safe.