Fixes pertaining to genesis initialization
The genesis init flags used to require more hand edits to the JSON
to get cardinal to accept it. After this, we still need to add the
hash and remove the blob schedule, but it's easier to get running.

I also realized that if someone queried the genesis block we were
replacing it with the latest block, so I adjusted the if statement
that handled that. It was originally intended to catch "latest" and
"pending" lookups.