Addresses a bug where consumer.Ready() wasn't firing in certain circumstances
Cardinal streams' consumer.Ready() is a bit finicky, and not terribly consistent
across streams implementations (kafka:// differs from null:// in important ways).

With the mux consumer, if you call Ready() multiple times you get different channels
and only one of them actually resolves when the stream is ready (probably the first one,
but that will depend on Go's scheduler).

Fixing this properly may have implications for other cardinal streams consumers,
which is a potentially big lift. Since we're aiming to get this out promptly for the upcoming
Fusaka hard fork (hitting Sepolia Tuesday), I elected to work around this in Cardinal.

So the update here is to get one instance of consumer.Ready(), and expect it to
only fire once. If it fires at line 115, we use the waiting variable to skip reading
on it at line 122.